ROP initial commit

+++ rebuild.sh
@@ -46,7 +46,7 @@ do
 	fi
 done
 
-for i in src/*/*.c
+for i in src/aleph_one/*.c
 do
 	doc=`date +%Y%m%d%H%M -r $i`
 	difference=$(( now - doc ))
@@ -55,6 +55,24 @@ do
 	    file_name="${i#src\/}"
 	    file_name="${file_name%.c}"
 	    command="gcc -m32 -w -static -ggdb -fno-stack-protector -z execstack $i -o bin/$file_name"
+		eval "rm -f bin/$file_name"
+	    if eval $command; then
+	    	echo "Created bin/$file_name"
+		else
+			echo "Failed to create bin/$file_name"
+		fi
+	fi
+done
+
+for i in src/rop/*.c
+do
+	doc=`date +%Y%m%d%H%M -r $i`
+	difference=$(( now - doc ))
+	if (( difference <= 15)); then
+	    file_name="$i"
+	    file_name="${i#src\/}"
+	    file_name="${file_name%.c}"
+	    command="gcc -m32 -w -ggdb -fno-stack-protector $i -o bin/$file_name"
 		eval "rm -f bin/$file_name"
 	    if eval $command; then
 	    	echo "Created bin/$file_name"
